Remove DialogHeader/DialogBody/DialogFooter and extract dialog styles
