+++ src/scripts/block-dialog/index.tsx
@@ -8,7 +8,6 @@ import * as punycode from "punycode/";
 import React, { useId, useMemo, useRef, useState } from "react";
 import { createRoot, type Root } from "react-dom/client";
 import icon from "../../icons/icon.svg";
-import dialogStyles from "../components/dialog.module.css";
 import { Dialog } from "../components/dialog.tsx";
 import { EmbeddedDialog } from "../components/embedded-dialog.tsx";
 import { MenuItem } from "../components/menu.tsx";
@@ -25,6 +24,7 @@ import { translate } from "../shared/locales.ts";
 import { getRegistrableDomain } from "../shared/registrable-domain.ts";
 import type { DialogTheme, MessageName0 } from "../shared/types.ts";
 import buttonStyles from "../styles/button.module.css";
+import dialogStyles from "../styles/dialog.module.css";
 import iconButtonStyles from "../styles/icon-button.module.css";
 import labelStyles from "../styles/label.module.css";
 import rowStyles from "../styles/row.module.css";
