+++ src/scripts/components/dialog.tsx
@@ -1,6 +1,7 @@
 import { Dialog as BaseDialog } from "@base-ui/react/dialog";
 import clsx from "clsx";
 import type React from "react";
+import sharedStyles from "../styles/dialog.module.css";
 import styles from "./dialog.module.css";
 
 export type DialogProps = React.JSX.IntrinsicElements["div"] & {
@@ -48,28 +49,13 @@ export function Dialog({
   );
 }
 
-export type DialogHeaderProps = React.JSX.IntrinsicElements["div"];
-
-export function DialogHeader({ className, ...props }: DialogHeaderProps) {
-  return <div {...props} className={clsx(styles.header, className)} />;
-}
-
 export type DialogTitleProps = React.JSX.IntrinsicElements["h2"];
 
 export function DialogTitle({ className, ...props }: DialogTitleProps) {
   return (
-    <BaseDialog.Title {...props} className={clsx(styles.title, className)} />
+    <BaseDialog.Title
+      {...props}
+      className={clsx(sharedStyles.title, className)}
+    />
   );
-}
-
-export type DialogBodyProps = React.JSX.IntrinsicElements["div"];
-
-export function DialogBody(props: DialogBodyProps) {
-  return <div {...props} />;
-}
-
-export type DialogFooterProps = React.JSX.IntrinsicElements["div"];
-
-export function DialogFooter({ className, ...props }: DialogFooterProps) {
-  return <div {...props} className={clsx(styles.footer, className)} />;
 }
