+++ src/scripts/options/general-section.tsx
@@ -3,13 +3,7 @@ import { Checkbox } from "@base-ui/react/checkbox";
 import openInNewSVG from "@mdi/svg/svg/open-in-new.svg";
 import clsx from "clsx";
 import { useEffect, useId, useState } from "react";
-import {
-  Dialog,
-  DialogBody,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "../components/dialog.tsx";
+import { Dialog, DialogTitle } from "../components/dialog.tsx";
 import { expandLinks } from "../components/link.tsx";
 import { Select, SelectOption } from "../components/select.tsx";
 import { SvgIcon } from "../components/svg-icon.tsx";
@@ -25,6 +19,7 @@ import {
 } from "../shared/utilities.ts";
 import buttonStyles from "../styles/button.module.css";
 import styles from "../styles/checkbox.module.css";
+import dialogStyles from "../styles/dialog.module.css";
 import iconButtonStyles from "../styles/icon-button.module.css";
 import indentStyles from "../styles/indent.module.css";
 import labelStyles from "../styles/label.module.css";
@@ -65,12 +60,12 @@ function ImportBlacklistForm({
 
   return (
     <>
-      <DialogHeader>
+      <div className={dialogStyles.header}>
         <DialogTitle>
           {translate("options_importBlacklistDialog_title")}
         </DialogTitle>
-      </DialogHeader>
-      <DialogBody>
+      </div>
+      <div>
         <div className={rowStyles.row}>
           <div className={rowStyles.rowItem}>
             <Select
@@ -138,8 +133,8 @@ function ImportBlacklistForm({
             </div>
           </div>
         </div>
-      </DialogBody>
-      <DialogFooter>
+      </div>
+      <div className={dialogStyles.footer}>
         <div className={clsx(rowStyles.row, rowStyles.right)}>
           <div className={rowStyles.rowItem}>
             <Button
@@ -186,7 +181,7 @@ function ImportBlacklistForm({
             )}
           </div>
         </div>
-      </DialogFooter>
+      </div>
     </>
   );
 }
