+++ src/scripts/options/subscription-section/add-dialog.tsx
@@ -2,19 +2,14 @@ import { Button } from "@base-ui/react/button";
 import { Input } from "@base-ui/react/input";
 import clsx from "clsx";
 import { useId, useState } from "react";
-import {
-  Dialog,
-  DialogBody,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "../../components/dialog.tsx";
+import { Dialog, DialogTitle } from "../../components/dialog.tsx";
 import { Select, SelectOption } from "../../components/select.tsx";
 import { translate } from "../../shared/locales.ts";
 import { sendMessage } from "../../shared/messages.ts";
 import { requestPermission } from "../../shared/permissions.ts";
 import type { Subscription, SubscriptionType } from "../../shared/types.ts";
 import buttonStyles from "../../styles/button.module.css";
+import dialogStyles from "../../styles/dialog.module.css";
 import inputStyles from "../../styles/input.module.css";
 import labelStyles from "../../styles/label.module.css";
 import rowStyles from "../../styles/row.module.css";
@@ -51,12 +46,12 @@ function AddForm({
 
   return (
     <>
-      <DialogHeader>
+      <div className={dialogStyles.header}>
         <DialogTitle>
           {translate("options_addSubscriptionDialog_title")}
         </DialogTitle>
-      </DialogHeader>
-      <DialogBody>
+      </div>
+      <div>
         <div className={rowStyles.row}>
           <div className={clsx(rowStyles.rowItem, rowStyles.expanded)}>
             <div className={clsx(labelStyles.wrapper, labelStyles.fullWidth)}>
@@ -133,8 +128,8 @@ function AddForm({
             />
           </div>
         </div>
-      </DialogBody>
-      <DialogFooter>
+      </div>
+      <div className={dialogStyles.footer}>
         <div className={clsx(rowStyles.row, rowStyles.right)}>
           <div className={rowStyles.rowItem}>
             <Button
@@ -170,7 +165,7 @@ function AddForm({
             </Button>
           </div>
         </div>
-      </DialogFooter>
+      </div>
     </>
   );
 }
