+++ src/scripts/options/subscription-section/rename-dialog.tsx
@@ -2,17 +2,12 @@ import { Button } from "@base-ui/react/button";
 import { Input } from "@base-ui/react/input";
 import clsx from "clsx";
 import { useId, useState } from "react";
-import {
-  Dialog,
-  DialogBody,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "../../components/dialog.tsx";
+import { Dialog, DialogTitle } from "../../components/dialog.tsx";
 import { translate } from "../../shared/locales.ts";
 import { sendMessage } from "../../shared/messages.ts";
 import type { Subscription, SubscriptionId } from "../../shared/types.ts";
 import buttonStyles from "../../styles/button.module.css";
+import dialogStyles from "../../styles/dialog.module.css";
 import inputStyles from "../../styles/input.module.css";
 import labelStyles from "../../styles/label.module.css";
 import rowStyles from "../../styles/row.module.css";
@@ -30,12 +25,12 @@ function RenameForm({
   const [name, setName] = useState(subscription.name);
   return (
     <>
-      <DialogHeader>
+      <div className={dialogStyles.header}>
         <DialogTitle>
           {translate("options_renameSubscriptionDialog_title")}
         </DialogTitle>
-      </DialogHeader>
-      <DialogBody>
+      </div>
+      <div>
         <div className={rowStyles.row}>
           <div className={clsx(rowStyles.rowItem, rowStyles.expanded)}>
             <div className={clsx(labelStyles.wrapper, labelStyles.fullWidth)}>
@@ -59,8 +54,8 @@ function RenameForm({
             />
           </div>
         </div>
-      </DialogBody>
-      <DialogFooter>
+      </div>
+      <div className={dialogStyles.footer}>
         <div className={clsx(rowStyles.row, rowStyles.right)}>
           <div className={rowStyles.rowItem}>
             <Button
@@ -82,7 +77,7 @@ function RenameForm({
             </Button>
           </div>
         </div>
-      </DialogFooter>
+      </div>
     </>
   );
 }
