+++ src/scripts/popup/serpinfo.tsx
@@ -4,7 +4,6 @@ import cog from "@mdi/svg/svg/cog.svg";
 import clsx from "clsx";
 import { useId, useRef, useState } from "react";
 import icon from "../../icons/icon.svg";
-import dialogStyles from "../components/dialog.module.css";
 import { EmbeddedDialog } from "../components/embedded-dialog.tsx";
 import { SvgIcon } from "../components/svg-icon.tsx";
 import { browser } from "../shared/browser.ts";
@@ -13,6 +12,7 @@ import { translate } from "../shared/locales.ts";
 import { sendMessage, sendMessageToTab } from "../shared/messages.ts";
 import { serpMatchesUrl } from "../shared/serpinfo-match.ts";
 import buttonStyles from "../styles/button.module.css";
+import dialogStyles from "../styles/dialog.module.css";
 import iconButtonStyles from "../styles/icon-button.module.css";
 import labelStyles from "../styles/label.module.css";
 import rowStyles from "../styles/row.module.css";
