+++ src/scripts/styles/dialog.module.css
@@ -1,0 +1,14 @@
+.header {
+  margin-bottom: 1em;
+}
+
+.title {
+  font-size: 1.125em;
+  font-weight: normal;
+  margin: 0;
+  overflow-wrap: break-word;
+}
+
+.footer {
+  margin-top: 2em;
+}
